+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/prettier/PrettierFormatStepTest.java
@@ -21,8 +21,8 @@ import java.io.IOException;
 
 import org.junit.jupiter.api.Test;
 
+import com.diffplug.spotless.ProcessRunner;
 import com.diffplug.spotless.maven.MavenIntegrationHarness;
-import com.diffplug.spotless.maven.MavenRunner.Result;
 import com.diffplug.spotless.maven.generic.Prettier;
 import com.diffplug.spotless.tag.NpmTest;
 
@@ -43,7 +43,7 @@ class PrettierFormatStepTest extends MavenIntegrationHarness {
 		return path;
 	}
 
-	private Result runExpectingError(String kind, String suffix) throws IOException, InterruptedException {
+	private ProcessRunner.Result runExpectingError(String kind, String suffix) throws IOException, InterruptedException {
 		String path = prepareRun(kind, suffix);
 		return mavenRunner().withArguments("spotless:apply").runHasError();
 	}
@@ -102,8 +102,8 @@ class PrettierFormatStepTest extends MavenIntegrationHarness {
 				"  <devDependencies><prettier>1.16.4</prettier></devDependencies>",
 				"</prettier>");
 
-		Result result = mavenRunner().withArguments("spotless:apply").runHasError();
-		assertThat(result.output()).contains(Prettier.ERROR_MESSAGE_ONLY_ONE_CONFIG);
+		ProcessRunner.Result result = mavenRunner().withArguments("spotless:apply").runHasError();
+		assertThat(result.stdOutUtf8()).contains(Prettier.ERROR_MESSAGE_ONLY_ONE_CONFIG);
 	}
 
 	@Test
@@ -156,8 +156,8 @@ class PrettierFormatStepTest extends MavenIntegrationHarness {
 				"  <prettierVersion>1.16.4</prettierVersion>",
 				"  <configFile>.prettierrc.yml</configFile>",
 				"</prettier>");
-		Result result = runExpectingError("typescript", suffix);
-		assertThat(result.output()).containsPattern("Running npm command.*npm install.* failed with exit code: 1");
+		ProcessRunner.Result result = runExpectingError("typescript", suffix);
+		assertThat(result.stdOutUtf8()).containsPattern("Running npm command.*npm install.* failed with exit code: 1");
 	}
 
 	@Test
@@ -174,7 +174,7 @@ class PrettierFormatStepTest extends MavenIntegrationHarness {
 				"  <configFile>.prettierrc.yml</configFile>",
 				"  <npmrc>${basedir}/.custom_npmrc</npmrc>",
 				"</prettier>");
-		Result result = runExpectingError("typescript", suffix);
-		assertThat(result.output()).containsPattern("Running npm command.*npm install.* failed with exit code: 1");
+		ProcessRunner.Result result = runExpectingError("typescript", suffix);
+		assertThat(result.stdOutUtf8()).containsPattern("Running npm command.*npm install.* failed with exit code: 1");
 	}
 }